// backend/src/types/character.ts
export enum CharacterRank {
  STONE_MONKEY = 'STONE_MONKEY',
  MONKEY_KING = 'MONKEY_KING',
  GREAT_SAGE = 'GREAT_SAGE',
  BUDDHA = 'BUDDHA',
  WUKONG = 'wukong'
}

export enum SubscriptionPlan {
  TRIAL = 'trial',
  STANDARD = 'standard',
  PRO = 'pro',
  NONE = 'none'
}

export enum StudyState {
  Studying = 'studying',
  Break = 'break',
  Finished = 'finished'
}

export enum FocusState {
  FOCUSED = 'focused',
  DISTRACTED = 'distracted',
  OFF_SEAT = 'off_seat'
}

export enum DistractedSubtype {
  PHONE = 'phone',
  TALK = 'talk',
  SLEEP = 'sleep',
  PLAY = 'play',
  ZONE = 'zone',
  DISTRACTED = 'distracted'
}